* issue_75-denyable-behavior: (21 commits)
  wrap NotPermittedException in PropelException
  corrected info.ymls
  when changing category_id or sort, roles for category must be given
  convenience methods defaults corrected
  re-built rights management with instance methods
  default all tables to deny by role_key
  allowed assigning rights to own roles
  tightened security around users
  implemented all types of owner_allowed configurations
  used alert log level
  correctly set the (as of now) useless option owner_allowed
  added denyable to documents and links
  added denyable to users (and related models)
  fixed static var Peer::$IGNORE_RIGHTS
  made Model-based permissions testable using the permission widget
  added translations
  added denyable to languages and strings
  added denyable to language_object_history
  Workaround for Propel issue 130
  added manual mayOperateOn method implementations
  ...

Closes #75 as fixed

+++ lib/model/om/BaseContentObject.php
@@ -659,6 +659,11 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 			$deleteQuery = ContentObjectQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(ContentObjectPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.backend_user", array("role_key" => "objects")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -702,6 +707,11 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(ContentObjectPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.backend_user", array("role_key" => "objects")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(ContentObjectPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -722,6 +732,11 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(ContentObjectPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.backend_user", array("role_key" => "objects")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(ContentObjectPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1938,6 +1953,26 @@ abstract class BaseContentObject extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(ContentObjectPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && ContentObjectPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return ContentObjectPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
