+++ lib/model/om/BaseDocumentCategory.php
@@ -682,6 +682,11 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 			if(ReferencePeer::hasReference($this)) {
 				throw new PropelException("Exception in ".__METHOD__.": tried removing an instance from the database even though it is still referenced.");
 			}
+			// denyable behavior
+			if(!(DocumentCategoryPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "documents")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -725,6 +730,11 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(DocumentCategoryPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "documents")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(DocumentCategoryPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -745,6 +755,11 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(DocumentCategoryPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "documents")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(DocumentCategoryPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1724,6 +1739,26 @@ abstract class BaseDocumentCategory extends BaseObject  implements Persistent
 	{
 		return ReferencePeer::getReferences($this);
 	}
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && DocumentCategoryPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return DocumentCategoryPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
+	}
+
 	// extended_timestampable behavior
 	
 	/**
