+++ lib/model/om/BaseDocumentType.php
@@ -590,6 +590,11 @@ abstract class BaseDocumentType extends BaseObject  implements Persistent
 			$deleteQuery = DocumentTypeQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(DocumentTypePeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "documents")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -633,6 +638,11 @@ abstract class BaseDocumentType extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(DocumentTypePeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "documents")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(DocumentTypePeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -653,6 +663,11 @@ abstract class BaseDocumentType extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(DocumentTypePeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "documents")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(DocumentTypePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1599,6 +1614,26 @@ abstract class BaseDocumentType extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(DocumentTypePeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && DocumentTypePeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return DocumentTypePeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
