+++ lib/model/om/BaseGroup.php
@@ -489,6 +489,11 @@ abstract class BaseGroup extends BaseObject  implements Persistent
 			$deleteQuery = GroupQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(GroupPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "users")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -532,6 +537,11 @@ abstract class BaseGroup extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(GroupPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "users")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(GroupPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -552,6 +562,11 @@ abstract class BaseGroup extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(GroupPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "users")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(GroupPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1653,6 +1668,26 @@ abstract class BaseGroup extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(GroupPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && GroupPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return GroupPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
