+++ lib/model/om/BaseGroupRole.php
@@ -501,6 +501,11 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 			$deleteQuery = GroupRoleQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(GroupRolePeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.by_role", array("role_key" => "users")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -544,6 +549,11 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(GroupRolePeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.by_role", array("role_key" => "users")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(GroupRolePeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -564,6 +574,11 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(GroupRolePeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.by_role", array("role_key" => "users")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(GroupRolePeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1319,6 +1334,26 @@ abstract class BaseGroupRole extends BaseObject  implements Persistent
 	public function __toString()
 	{
 		return (string) $this->exportTo(GroupRolePeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && GroupRolePeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return GroupRolePeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
