+++ lib/model/om/BaseLanguageObjectHistory.php
@@ -584,6 +584,11 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 			$deleteQuery = LanguageObjectHistoryQuery::create()
 				->filterByPrimaryKey($this->getPrimaryKey());
 			$ret = $this->preDelete($con);
+			// denyable behavior
+			if(!(LanguageObjectHistoryPeer::isIgnoringRights() || $this->mayOperate("delete"))) {
+				throw new PropelException(new NotPermittedException("delete.backend_user", array("role_key" => "language_object_history")));
+			}
+
 			if ($ret) {
 				$deleteQuery->delete($con);
 				$this->postDelete($con);
@@ -627,6 +632,11 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 			$ret = $this->preSave($con);
 			if ($isInsert) {
 				$ret = $ret && $this->preInsert($con);
+				// denyable behavior
+				if(!(LanguageObjectHistoryPeer::isIgnoringRights() || $this->mayOperate("insert"))) {
+					throw new PropelException(new NotPermittedException("insert.backend_user", array("role_key" => "language_object_history")));
+				}
+
 				// extended_timestampable behavior
 				if (!$this->isColumnModified(LanguageObjectHistoryPeer::CREATED_AT)) {
 					$this->setCreatedAt(time());
@@ -647,6 +657,11 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 
 			} else {
 				$ret = $ret && $this->preUpdate($con);
+				// denyable behavior
+				if(!(LanguageObjectHistoryPeer::isIgnoringRights() || $this->mayOperate("update"))) {
+					throw new PropelException(new NotPermittedException("update.backend_user", array("role_key" => "language_object_history")));
+				}
+
 				// extended_timestampable behavior
 				if ($this->isModified() && !$this->isColumnModified(LanguageObjectHistoryPeer::UPDATED_AT)) {
 					$this->setUpdatedAt(time());
@@ -1432,6 +1447,26 @@ abstract class BaseLanguageObjectHistory extends BaseObject  implements Persiste
 	public function __toString()
 	{
 		return (string) $this->exportTo(LanguageObjectHistoryPeer::DEFAULT_STRING_FORMAT);
+	}
+
+	// denyable behavior
+	public function mayOperate($sOperation, $oUser = false) {
+		if($oUser === false) {
+			$oUser = Session::getSession()->getUser();
+		}
+		if($oUser && ($this->isNew() || $this->getCreatedBy() === $oUser->getId()) && LanguageObjectHistoryPeer::mayOperateOnOwn($oUser, $this, $sOperation)) {
+			return true;
+		}
+		return LanguageObjectHistoryPeer::mayOperateOn($oUser, $this, $sOperation);
+	}
+	public function mayBeInserted($oUser = false) {
+		return $this->mayOperate($oUser, "insert");
+	}
+	public function mayBeUpdated($oUser = false) {
+		return $this->mayOperate($oUser, "update");
+	}
+	public function mayBeDeleted($oUser = false) {
+		return $this->mayOperate($oUser, "delete");
 	}
 
 	// extended_timestampable behavior
